[배인수] 1. 서버 Teleport Behavior 구현 2. 서버 JSONHelper 소스 다듬기 3. 서버 프로젝트 추가 포함 디렉토리 수정 4. 패킷 헤더 내 사이즈 관련 버그 수정

+++ Client_Root/Client/Assets/Scripts/Network/Network.cs
@@ -234,7 +234,7 @@ public class Network : MonoSingleton<Network>
 		try
 		{
             byte[] byteSerializedData = msg.Serialize();
-            int nTotalSize = NetworkDefines.MESSAGE_HEADER_SIZE + byteSerializedData.Length;		//	msg id(2) + msg length info(2)
+            ushort nTotalSize = (ushort)(NetworkDefines.MESSAGE_HEADER_SIZE + byteSerializedData.Length);		//	msg id(2) + msg length info(2)
 			byte[] byteData = new byte[nTotalSize];
 
 			byteData [0] = BitConverter.GetBytes(msg.GetID())[1];
